feat: Enhance security and authentication in media and products services, update product model, and refactor related components

+++ products-service/products/src/main/java/com/example/products/controllers/ProductController.java
@@ -36,7 +36,7 @@ public class ProductController {
         this.productService = productService;
     }
 
-    @GetMapping({ "", "/" })
+    @GetMapping("/")
     @PermitAll
     public ResponseEntity<ApiResponse<List<Product>>> getProducts() {
         return ResponseEntity.ok(ApiResponse.success(productService.getAllProducts()));
@@ -57,14 +57,12 @@ public class ProductController {
         return ResponseEntity.ok(ApiResponse.success(product));
     }
 
-    @PostMapping({ "", "/" })
-    public ResponseEntity<ApiResponse<Product>> createProduct(@RequestBody @Valid CreateProdutDto productDto,
+    @PostMapping("/")
+    public ResponseEntity<ApiResponse<Product>> createProduct(
+            @RequestBody @Valid CreateProdutDto productDto,
             Authentication authentication) {
-        /*******************************************/
-        /* TODO: Need Owner ID from authentication */
-        /*******************************************/
 
-        UUID userId = UUID.randomUUID();
+        String userId = (String) authentication.getPrincipal();
 
         Product createdProduct = this.productService.createProduct(productDto, userId);
 
@@ -81,16 +79,14 @@ public class ProductController {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .body(ApiResponse.error("Product not found", HttpStatus.NOT_FOUND));
         }
-        /*******************************************/
-        /* TODO: Need Owner ID from authentication */
-        /*******************************************/
-        UUID userId = UUID.randomUUID();
 
-        // if (!product.getUserId().equals(userId)) {
-        // return ResponseEntity.status(HttpStatus.FORBIDDEN)
-        // .body(ApiResponse.error("You are not the owner of this product",
-        // HttpStatus.FORBIDDEN));
-        // }
+        String userId = (String) authentication.getPrincipal();
+
+        if (!product.getUserId().equals(userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(ApiResponse.error("You are not the owner of this product",
+                            HttpStatus.FORBIDDEN));
+        }
 
         this.productService.deleteProduct(id);
 
@@ -109,16 +105,14 @@ public class ProductController {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .body(ApiResponse.error("Product not found", HttpStatus.NOT_FOUND));
         }
-        /*******************************************/
-        /* TODO: Need Owner ID from authentication */
-        /*******************************************/
-        UUID userId = UUID.randomUUID();
 
-        // if (!product.getUserId().equals(userId)) {
-        // return ResponseEntity.status(HttpStatus.FORBIDDEN)
-        // .body(ApiResponse.error("You are not the owner of this product",
-        // HttpStatus.FORBIDDEN));
-        // }
+        String userId = (String) authentication.getPrincipal();
+
+        if (!product.getUserId().equals(userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(ApiResponse.error("You are not the owner of this product",
+                            HttpStatus.FORBIDDEN));
+        }
 
         Product updateProduct = this.productService.updateProduct(product, productDto);
